pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-agentic-rag"
version = "0.1.0"
description = "Production-quality local agentic RAG starter with hybrid retrieval, citations, Streamlit, and MCP."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "httpx>=0.28.1",
  "numpy>=2.2.0",
  "pypdf>=5.3.0",
  "python-docx>=1.1.2",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
  "typer>=0.16.0",
]

[project.optional-dependencies]
faiss = [
  "faiss-cpu>=1.10.0; python_version < '3.14' and platform_system != 'Windows'"
]
ui = [
  "streamlit>=1.44.0"
]
web = [
  "starlette>=0.46.2",
  "uvicorn>=0.35.0"
]
mcp = [
  "mcp>=1.7.1"
]
dev = [
  "pytest>=8.3.5",
  "pytest-cov>=6.0.0",
  "reportlab>=4.4.0"
]

[project.scripts]
local-rag = "local_agentic_rag.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
